[SYCL-MLIR] Simplify polygeist and sycl conversion passes
This includes:

- Having a single <PassName>.h and <PassName>.cpp file for each pass;
- Declare passes code in the mlir namespace, enabling use of
  auto-generated factory functions;
- Use pass-specific options for configuration instead of
  LowerToLLVMOptions and letting the passes fill those options as
  needed;
- Drop option controlling "bare pointer conversion" and asserting it
  is set when calling function populating patterns, as our passes rely
  on such option being passed, making it "not an option".
