+++ mlir-sycl/lib/Conversion/SYCLToGPU/CMakeLists.txt
@@ -1,6 +1,5 @@
 add_mlir_conversion_library(MLIRSYCLToGPU
   SYCLToGPU.cpp
-  SYCLToGPUPass.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${MLIR_SYCL_SOURCE_DIR}/mlir/Conversion/SYCLToGPU
