+++ mlir-sycl/lib/Conversion/SYCLToLLVM/CMakeLists.txt
@@ -2,7 +2,6 @@ add_mlir_conversion_library(MLIRSYCLToLLVM
   DialectBuilder.cpp
   SYCLFuncRegistry.cpp
   SYCLToLLVM.cpp
-  SYCLToLLVMPass.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${MLIR_SYCL_SOURCE_DIR}/mlir/Conversion/SYCLToLLVM
