--- mlir-sycl/lib/Conversion/SYCLToLLVM/SYCLToLLVMPass.cpp
@@ -1,91 +1,0 @@
-//===- SYCLToLLVMPass.cpp - SYCL to LLVM Passes ---------------------------===//
-//
-// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-//
-//===----------------------------------------------------------------------===//
-//
-// This file implements a pass to convert MLIR SYCL ops into LLVM ops
-//
-//===----------------------------------------------------------------------===//
-
-#include "mlir/Conversion/SYCLToLLVM/SYCLToLLVMPass.h"
-
-#include "mlir/Conversion/FuncToLLVM/ConvertFuncToLLVM.h"
-#include "mlir/Conversion/LLVMCommon/LoweringOptions.h"
-#include "mlir/Conversion/LLVMCommon/TypeConverter.h"
-#include "mlir/Conversion/SYCLToLLVM/SYCLToLLVM.h"
-#include "mlir/Dialect/Arith/IR/Arith.h"
-#include "mlir/Dialect/Func/Transforms/FuncConversions.h"
-#include "mlir/Dialect/LLVMIR/LLVMDialect.h"
-#include "mlir/Dialect/Polygeist/Transforms/Passes.h"
-#include "mlir/Dialect/SYCL/IR/SYCLOps.h"
-#include "mlir/Dialect/SYCL/IR/SYCLOpsDialect.h"
-
-using namespace mlir;
-
-namespace mlir {
-#define GEN_PASS_DEF_CONVERTSYCLTOLLVM
-#include "mlir/Conversion/SYCLPasses.h.inc"
-#undef GEN_PASS_DEF_CONVERTSYCLTOLLVM
-} // namespace mlir
-
-namespace {
-/// A pass converting MLIR SYCL operations into LLVM dialect.
-class ConvertSYCLToLLVMPass
-    : public impl::ConvertSYCLToLLVMBase<ConvertSYCLToLLVMPass> {
-public:
-  ConvertSYCLToLLVMPass() = default;
-
-  explicit ConvertSYCLToLLVMPass(bool useBarePtrCallConv) {
-    this->useBarePtrCallConv = useBarePtrCallConv;
-  }
-
-  void runOnOperation() override;
-};
-} // namespace
-
-void ConvertSYCLToLLVMPass::runOnOperation() {
-  MLIRContext *context = &getContext();
-  ModuleOp module = getOperation();
-
-  LowerToLLVMOptions options(&getContext());
-  options.useBarePtrCallConv = useBarePtrCallConv;
-  LLVMTypeConverter converter(&getContext(), options);
-
-  RewritePatternSet patterns(context);
-
-  if (useBarePtrCallConv) {
-    // Keep these at the top; these should be run before the rest of
-    // function conversion patterns.
-    populateReturnOpTypeConversionPattern(patterns, converter);
-    populateCallOpTypeConversionPattern(patterns, converter);
-    populateAnyFunctionOpInterfaceTypeConversionPattern(patterns, converter);
-    polygeist::populateBareMemRefToLLVMConversionPatterns(converter, patterns);
-  }
-
-  sycl::populateSYCLToLLVMConversionPatterns(converter, patterns);
-  populateFuncToLLVMConversionPatterns(converter, patterns);
-
-  ConversionTarget target(*context);
-  target.addIllegalDialect<sycl::SYCLDialect>();
-  // TODO: connect wires to lower this operation.
-  target.addLegalOp<sycl::SYCLLocalIDOp>();
-  target.addLegalDialect<LLVM::LLVMDialect>();
-  target.addLegalDialect<arith::ArithDialect>();
-
-  target.addLegalOp<ModuleOp>();
-  if (failed(applyPartialConversion(module, target, std::move(patterns))))
-    signalPassFailure();
-}
-
-std::unique_ptr<OperationPass<ModuleOp>>
-mlir::sycl::createConvertSYCLToLLVMPass(const LowerToLLVMOptions &options) {
-  return std::make_unique<ConvertSYCLToLLVMPass>(options.useBarePtrCallConv);
-}
-
-std::unique_ptr<OperationPass<ModuleOp>>
-mlir::sycl::createConvertSYCLToLLVMPass() {
-  return std::make_unique<ConvertSYCLToLLVMPass>();
-}
