+++ mlir-sycl/lib/Conversion/SYCLToSPIRV/CMakeLists.txt
@@ -1,6 +1,5 @@
 add_mlir_conversion_library(MLIRSYCLToSPIRV
   SYCLToSPIRV.cpp
-  SYCLToSPIRVPass.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${MLIR_SYCL_SOURCE_DIR}/mlir/Conversion/SYCLToSPIRV
