+++ polygeist/test/polygeist-opt/bareptrlowering.mlir
@@ -1,4 +1,4 @@
-// RUN: polygeist-opt %s --convert-polygeist-to-llvm="use-bare-ptr-memref-call-conv" --split-input-file | FileCheck %s
+// RUN: polygeist-opt %s --convert-polygeist-to-llvm --split-input-file | FileCheck %s
 
 // CHECK-LABEL:   llvm.func @ptr_ret_static(i64) -> !llvm.ptr<i64>
